Done- algorithm calculate discount

+++ wsep192/src/Domain/Product.cs
@@ -12,10 +12,10 @@ namespace src.Domain
         private String productName;
         private String category;
         private String details;
-        private int price;
+        private double price;
         private int productRate;
 
-        public Product(int id, string productName, string category, string details, int price)
+        public Product(int id, string productName, string category, string details, double price)
         {
             this.Id = id;
             this.ProductName = productName;
@@ -29,7 +29,7 @@ namespace src.Domain
         public string ProductName { get => productName; set => productName = value; }
         public string Category { get => category; set => category = value; }
         public string Details { get => details; set => details = value; }
-        public int Price { get => price; set => price = value; }
+        public double Price { get => price; set => price = value; }
         public int ProductRate { get => productRate; set => productRate = value; }
 
         public bool compareProduct(Filter filter)
